test: tidy up lexer tests
